added tool tip with auto sizing not working
fuck this auto size bullshit not working its so lame

+++ Assets/Scrripts/ItemScripts/Curio.cs
@@ -6,6 +6,6 @@ using UnityEngine;
 public class Curio : Item
 {
     public CurioEffect curiEef;
-    public float CurioNum;
+    public int CurioNum;
 }
 public enum CurioEffect { Koban,healing,heal,incrArea,revive,greed }
